requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.0.1
anytree==2.9.0
APScheduler==3.10.4
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
branca==0.6.0
bs4==0.0.1
treelib
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
et-xmlfile==1.1.0
Fiona==1.9.4.post1
Flask==2.3.2
folium==0.14.0
fonttools==4.42.1
frozenlist==1.3.3
geopandas==0.13.2
gitdb==4.0.10
GitPython==3.1.32
graphviz==0.20.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
plotly==5.16.1
protobuf==4.24.1
pyarrow==12.0.1
pydeck==0.8.0
Pygments==2.16.1
PyJWT==2.7.0
Pympler==1.0.1
pyparsing==3.0.9
pyproj==3.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==1.4.44
streamlit
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
twilio==8.4.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.21.2
waitress==2.1.2
watchdog==3.0.0
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
networkx